add tests folder
add Descuento model test

add detalle venta test

add devolucion test

add direccion model test

add generico model test

add marca model test

add medio pago model test

add motivo model test

add direccion model test

add otro motivo model test

add pais model test

add producto model test

add region provincia model test

fix rol model test

add usuario model test

add venta model test

+++ src/main/java/com/aespsoftware/pngcommerce/models/Descuento.java
@@ -23,7 +23,7 @@ public class Descuento {
 	private Producto producto;
 	
 	@Column(name = "porcentaje_dscto", nullable = false)
-	private int porcentajeDscto;
+	private double porcentajeDscto;
 	
 	@Column(name = "activo", nullable = false)
 	private boolean activo;
